separated spin script from functions
