+++ NS_vs_MCMC/NS_MCMC_spin_comparison_violin_plots.py
@@ -5,287 +5,12 @@ import datetime
 import pandas as pd
 import seaborn as sns
 import matplotlib.patches as mpatches
+from MCMC_NS_comparisons_violin_plots import get_params, extract_MCMC_data, extract_NS_data, combine_NS_MCMC_arrays, violin_plotter
 
 #TIMER
 start_timer = datetime.datetime.now()
 print(start_timer)
 print('\n')
-
-#-----------------------------------------
-#FUNCTIONS
-
-#CHECK WHICH PARAMETERS ARE NEEDED FOR PLOTTING
-def check_params_present(MCMC_file_location, MCMC_base_filenames):
-    
-    #ALL FITS FILES SHOULD HAVE THE SAME FIELDS
-    hdul = fits.open('%s/%s1.fits' %(MCMC_file_location, MCMC_base_filenames))
-    cols = hdul[1].columns
-    data = hdul[1].data
-    
-    checked_params = []
-    
-    #TEST FOR N_O_III AND N_NE_III
-    try:
-        OIII_array = data['O_III__12']
-        NeIII_array = data['Ne_III__15']
-        print("N_O_III AND N_NE_III INCL.")
-        checked_params.extend([['O_III__12'], ['Ne_III__15']])
-    except Exception:
-        checked_params.extend(['NO_DATA','NO_DATA'])
-        
-    #TEST IF SPIN FREE
-    try:
-        a_array = data["a__37"]
-        print("SPIN FREE")
-        checked_params.extend([["a__37"]])
-    except Exception:
-        checked_params.extend(['NO_DATA'])
-        
-    #CHECK FOR GABS
-    try:
-        LineE_array = data['LineE__61']
-        Strength_array1 = data['Strength__63']
-        Strength_array2 = data['Strength__66']
-        Strength_array3 = data['Strength__69']
-        checked_params.extend([['LineE__61', 'Strength__63', 'Strength__66', 'Strength__69']])
-        print("GABS INCL.")
-    except Exception:
-        checked_params.extend(['NO_DATA'])
-        
-    #CHECK FOR SMEDGE
-    try:
-        Smedge_E_array = data['edgeE__73']
-        Smedge_Tau_array = data['MaxTau__74']
-        Smedge_w_array = data['width__76']
-        print("SMEDGE INCL.")
-        checked_params.extend([['edgeE__73', 'MaxTau__74', 'width__76']])
-    except Exception:
-        checked_params.extend(['NO_DATA'])
-        
-    #CHECK FOR NUSTAR DATA
-    try:
-        factor_1_array = data['factor__61']
-        factor_2_array = data['factor__121']
-        checked_params.extend([['factor__61', 'factor__121']])
-        print("NuSTAR DATA INCL.")
-    except Exception:
-        print("ONLY XMM DATA")
-        checked_params.extend(['NO_DATA'])
-        
-    return checked_params
-
-#COLLATES THE LIST OF ALL PARAMETERS FOR CHECK FITS FILES
-def get_params(MCMC_file_location, MCMC_base_filenames, start_timer):
-    
-    #PARAMETERS THAT MIGHT BE MISSING
-    checked_params = check_params_present(MCMC_file_location, MCMC_base_filenames)
-    
-    #PARAMETERS THAT ARE ALWAYS PRESENT
-    standard_params = [
-        ["H__2", "O_I__10", "O_II__11"],
-        ["Ne_I__13","Ne_II__14"],
-        [ "Fe__31","Gamma__33","FracSctr__34"],
-        ["i__38","Mbh__39","Mdd__40","Dbh__41", "hd__42", 
-         "Index1__51", "Index2__52", "gamma__54",
-         "logxi__55", "Afe__57", "norm__60"],
-         ]
-    
-    #SORT PARAMETERS INTO THE CORRECT ORDER
-    params = []
-    for n in range(len(checked_params)):
-        try:
-            params.extend(standard_params[n])
-        except Exception:
-            print("RAN OUT OF STANDARD PARAMS, CONTINUING WITH CHECKED ONLY")
-        if (checked_params[n] != 'NO_DATA'):
-            params.extend(checked_params[n])
-
-    print("\nPARAM LIST MADE (LENGTH: %s PARAMS), TIME:" %len(params), datetime.datetime.now() - start_timer)  
-    print("--------------------------------------------\n") 
-            
-    return params
-
-#EXTRACTS THE DATA FROM EACH OF THE MCMC SAMPLES AND PUTS INTO AN ARRAY PER PARAMETER
-def extract_MCMC_data(MCMC_file_location, MCMC_base_filenames, params, start_timer, no_walkers, scale_down_factor):
-
-    i = 0
-    MCMC_values = [[]] * len(params) #EMPTY ARRAY FOR ALL MCMC PARAMETER VALUES
-    
-    #EXTRACTS DATA FOR EACH FIELD AND EACH SAMPLE AND ADDS TO THE ARRAY
-    while (i < len(params)):
-        for n in range(3):   
-            hdul = fits.open('%s/%s%s.fits' %(MCMC_file_location, MCMC_base_filenames, n+1))
-            cols = hdul[1].columns
-            data = hdul[1].data
-            
-            param_array = data[params[i]]
-            
-            if (n == 0):
-                MCMC_values[i] = param_array
-            else:
-                MCMC_values[i] = np.concatenate((MCMC_values[i], param_array))
-                
-        print('PARAM %s: %s, TIME: %s' %(i+1, params[i], datetime.datetime.now() - start_timer))
-        i = i + 1
-
-    print("\nMCMC DATA EXTRACTED")  
-    print("--------------------------------------------")
-    
-    thinned_MCMC_values = MCMC_thinning(params, MCMC_values, no_walkers, scale_down_factor)
-    
-    return thinned_MCMC_values
-
-#THINS THE MCMC CHAIN PER WALKER
-def MCMC_thinning(params, MCMC_values, no_walkers, scale_down_factor):
-    thinned_MCMC_values = [[]] * len(params)
-    
-    for n in range(len(params)):
-        for k in range(no_walkers):
-            param_values_thinned = MCMC_values[n][k::scale_down_factor]
-            
-            if (n == 0) & (k == 0):
-                thinned_MCMC_values[n] = param_values_thinned
-            else:
-                thinned_MCMC_values[n] = np.concatenate((thinned_MCMC_values[n], param_values_thinned))
-                
-        print('PARAM %s: %s, TIME: %s' %(n + 1, params[n], datetime.datetime.now() - start_timer))
-    print("\nMCMC DATA THINNED")  
-    print("--------------------------------------------")
-    
-    return thinned_MCMC_values
-
-#EXTRACTS THE DATA FROM THE NS FILE AND PUTS INTO AN ARRAY PER PARAMETER
-def extract_NS_data(NS_file_location, NS_filename, params):
-    Nested_Sampling = open("%s/%s.txt" %(NS_file_location, NS_filename), "r")
-    lines = Nested_Sampling.readlines()
-    Nested_Sampling.close()
-    
-    #IGNORE FIRST LINE AND SORT VALUES INTO THE ARRAY
-    print("IGNORED FIRST LINE:", lines[0])
-    NS_values = [[] for _ in range(len(params))] #EMPTY ARRAY FOR ALL NS PARAMETER VALUES
-    for line in lines[1:]:  # skip first line
-        values = line.split()
-        for n in range(len(params)):
-            NS_values[n].append(float(values[n]))
-
-    NS_values = unlog_NS_Mdd_and_norm_R(params, NS_values, lines)
-
-    print("\nNS DATA EXTRACTED")  
-    print("--------------------------------------------\n")
-
-    return NS_values
-
-#UNLOGS NS PARAMETERS TO BRING IN LINE WITH MCMC VALUES
-def unlog_NS_Mdd_and_norm_R(params, NS_values, lines):
-    #CHECK WHICH POSITIONS Mdd & norm_R ARE
-    Mdd_pos = np.where(np.array(params) == "Mdd__40")[0][0]
-    norm_R_pos = np.where(np.array(params) == "norm__60")[0][0]
-    
-    #UNLOG PARAMETERS
-    for n in range(len(NS_values[0])):
-        NS_values[Mdd_pos][n] = 10**float(NS_values[Mdd_pos][n])
-        NS_values[norm_R_pos][n] = 10**float(NS_values[norm_R_pos][n])
-
-    return NS_values
-
-def combine_NS_MCMC_arrays(MCMC_values, NS_values, params, diff):
-    combined_values = np.array([None]*(len(params)+2))
-    diff_values = np.array([diff]*(len(MCMC_values[0]) + len(NS_values[0])))
-    NS_MCMC_values = np.array(["MCMC"]*len(MCMC_values[0]) + ["NS"]*len(NS_values[0]))
-    
-    for n in range(len(params)):
-        combined_values[n] = np.concatenate((MCMC_values[n], NS_values[n]), axis=None)
-    combined_values[-2] = diff_values
-    combined_values[-1] = NS_MCMC_values
-    
-    return combined_values
-
-def violin_plotter(combined_diff_values, params, diff, x_axis_label, save_filename = False, plot_in_console = False):
-    
-    pink_colour = "#DC267F"
-    blue_colour = "#648FFF"
-    
-    dfs = []
-    
-    #PUT VALUES INTO DATAFRAME WITH ADDITIONAL PARAMS FOR PLOTTING
-    dfs = []
-    for diff_group in combined_diff_values:
-        param_arrays = diff_group[:-2]
-        diff_array = diff_group[-2]
-        label_array = diff_group[-1]
-        
-        #BUILD DICT DATA FRAME
-        col_dict = {params[i]: param_arrays[i] for i in range(len(params))}
-        col_dict[diff] = diff_array
-        col_dict['NS or MCMC'] = label_array
-        
-        df_group = pd.DataFrame(col_dict)
-        dfs.append(df_group)
-    
-    #COMBINE ALL DIFF GROUPS
-    df = pd.concat(dfs, ignore_index=True)
-    
-    #PLOTTING SET UP
-    columns = 4
-    rows = int(np.ceil(len(params)/4))
-    
-    fig = plt.figure(figsize=(6*columns*len(combined_diff_values), 6*rows), constrained_layout=True)
-    grid = fig.add_gridspec(rows, columns)
-    
-    label_array = []
-    
-    for n in range(len(params)):        
-        grid_row = int(n/columns)
-        grid_column = n%columns
-        ax = fig.add_subplot(grid[grid_row, grid_column])
-    
-        temp = sns.violinplot(data=df, x=diff, y=params[n], 
-                       hue='NS or MCMC', 
-                       split=True,
-                       palette=[blue_colour,pink_colour], 
-                       cut = 0, bw=.2, inner = None
-                       )
-        
-        label_array.append(temp)
-        ax.get_legend().remove()
-        
-        #ADD LEGEND BELOW THE FINAL PLOT
-        mcmc_patch = mpatches.Patch(color=blue_colour, label="MCMC")
-        ns_patch = mpatches.Patch(color=pink_colour, label="Nested Sampling")
-        fig.legend(handles=[mcmc_patch, ns_patch], loc="lower center", fontsize=30, 
-                   ncol=2, bbox_to_anchor=(0.5, 0), frameon=False)
-        
-        #PLOT SETTINGS
-        plt.xlabel(x_axis_label, size=34, labelpad = 10)
-        plt.ylabel(params[n], size=34, labelpad = 15)
-        plt.yticks(fontsize=26)
-        plt.xticks(fontsize=26)
-        
-        plt.tick_params(axis="both", direction = "in", width = 1.5, length = 4)
-        
-        sns.despine()
-        sns.set_style("ticks")
-        ax.spines['left'].set_linewidth(1.5)
-        ax.spines['bottom'].set_linewidth(1.5)
-        
-        print("PLOTTING TIME: ", params[n], ":", datetime.datetime.now() - start_timer)
-     
-    #ADJUST SUBPLOT SPACING
-    wspace = 1 / len(combined_diff_values)
-    plt.subplots_adjust(wspace = wspace, hspace = 0.5)
-    
-    #SAVE PLOT TO A PNG FILE
-    if (save_filename!= False):
-        save_timer = datetime.datetime.now() #ADDITIONAL TIMER FOR PLOT SAVING
-        plt.savefig("%s.png" %save_filename, bbox_inches = "tight")
-        print("\nSAVE TIME: %s" %(datetime.datetime.now() - save_timer))
-    
-    #PLOT IN CONSOLE
-    if (plot_in_console):
-        plt.show()
-    
-#-----------------------------------------
-#MAIN EXAMPLE SCRIPT
 
 #ALL MCMC FILES
 all_MCMC_base_filenames = [
@@ -358,7 +83,7 @@ for n in range(len(all_MCMC_base_filenames)):
     all_combined_values.append(combined_values)
 
 #MAKE THE VIOLIN PLOTS
-violin_plotter(all_combined_values, all_params[0], "REV", "spin", save_filename = "rev3533_spins_MCMC_NS_Violin_Comparison", plot_in_console = True)
+violin_plotter(all_combined_values, all_params[0], "REV", "spin", start_timer, 
+               save_filename = "rev3533_spins_MCMC_NS_Violin_Comparison", plot_in_console = True)
 
 print("\nTOTAL TIME: ", datetime.datetime.now() - start_timer)
-
